Refactor: MSW - StoryBook 로직 개선
