+++ .storybook/preview.js
@@ -2,6 +2,7 @@ import { withMsw } from './mswDecorator';
 import { queryClient } from '../modules/app/api/queryClient';
 import QueryClientProvider from '@shared/provider/QueryClientProvider';
 import FontThemeProvider from '@shared/provider/FontProvider';
+import handlers from '../modules/app/mocks/handlers';
 
 export const decorators = [
   withMsw,
@@ -18,3 +19,9 @@ export const decorators = [
     </QueryClientProvider>
   ),
 ];
+
+export const parameters = {
+  msw: {
+    handlers: [...handlers],
+  },
+};
