+++ jest/mock.setup.ts
@@ -10,15 +10,3 @@ jest.mock('react-native-maps', () => {
 
 jest.mock('expo-font');
 jest.mock('expo-asset');
-
-jest.mock('@react-navigation/native', () => {
-  const actualNav: unknown[] = jest.requireActual('@react-navigation/native');
-
-  return {
-    ...actualNav,
-    useNavigation: () => ({
-      navigate: jest.fn(),
-      dispatch: jest.fn(),
-    }),
-  };
-});
